+++ validator_derive/src/tokens/contains.rs
@@ -1,12 +1,11 @@
 use quote::quote;
-use syn::Ident;
 
 use crate::types::Contains;
 use crate::utils::{quote_code, quote_message};
 
 pub fn contains_tokens(
     contains: Contains,
-    field_name: &Ident,
+    field_name: &proc_macro2::TokenStream,
     field_name_str: &str,
 ) -> proc_macro2::TokenStream {
     let p = contains.pattern;
@@ -17,11 +16,11 @@ pub fn contains_tokens(
     let code = quote_code(contains.code, "contains");
 
     quote! {
-        if !self.#field_name.validate_contains(#needle) {
+        if !#field_name.validate_contains(#needle) {
             #code
             #message
             #needle_err
-            err.add_param(::std::borrow::Cow::from("value"), &self.#field_name);
+            err.add_param(::std::borrow::Cow::from("value"), &#field_name);
             errors.add(#field_name_str, err);
         }
     }