+++ validator_derive/src/tokens/mod.rs
@@ -11,6 +11,5 @@ pub mod non_control_character;
 pub mod range;
 pub mod regex;
 pub mod required;
-pub mod required_nested;
 pub mod schema;
 pub mod url;