+++ validator_derive/src/tokens/must_match.rs
@@ -1,12 +1,11 @@
 use quote::quote;
-use syn::Ident;
 
 use crate::types::MustMatch;
 use crate::utils::{quote_code, quote_message};
 
 pub fn must_match_tokens(
     must_match: MustMatch,
-    field_name: &Ident,
+    field_name: &proc_macro2::TokenStream,
     field_name_str: &str,
 ) -> proc_macro2::TokenStream {
     let o = must_match.other;
@@ -17,11 +16,11 @@ pub fn must_match_tokens(
     let code = quote_code(must_match.code, "must_match");
 
     quote! {
-        if !::validator::validate_must_match(&self.#field_name, &#other) {
+        if !::validator::validate_must_match(&#field_name, &#other) {
             #code
             #message
             #other_err
-            err.add_param(::std::borrow::Cow::from("value"), &self.#field_name);
+            err.add_param(::std::borrow::Cow::from("value"), &#field_name);
             errors.add(#field_name_str, err);
         }
     }