+++ validator_derive/src/tokens/regex.rs
@@ -1,12 +1,11 @@
 use quote::quote;
-use syn::Ident;
 
 use crate::types::Regex;
 use crate::utils::{quote_code, quote_message};
 
 pub fn regex_tokens(
     regex: Regex,
-    field_name: &Ident,
+    field_name: &proc_macro2::TokenStream,
     field_name_str: &str,
 ) -> proc_macro2::TokenStream {
     let path = regex.path;
@@ -14,10 +13,10 @@ pub fn regex_tokens(
     let code = quote_code(regex.code, "regex");
 
     quote! {
-        if !&self.#field_name.validate_regex(&#path) {
+        if !&#field_name.validate_regex(&#path) {
             #code
             #message
-            err.add_param(::std::borrow::Cow::from("value"), &self.#field_name);
+            err.add_param(::std::borrow::Cow::from("value"), &#field_name);
             errors.add(#field_name_str, err);
         }
     }