+++ validator_derive_tests/tests/run-pass/custom.rs
@@ -1,3 +1,4 @@
+use std::borrow::Cow;
 use validator::{Validate, ValidationError};
 
 #[derive(Validate)]
@@ -10,9 +11,17 @@ struct Test {
 struct TestPath {
     #[validate(custom(function = "crate::validate_something"))]
     s: String,
+    #[validate(custom(function = "crate::validate_number"))]
+    n: u8,
+    #[validate(custom(function = "crate::validate_something"))]
+    r: Cow<'static, str>
 }
 
 fn validate_something(_s: &str) -> Result<(), ValidationError> {
+    Ok(())
+}
+
+fn validate_number(_s: u8) -> Result<(), ValidationError> {
     Ok(())
 }