+++ validator_derive_tests/tests/run-pass/default_struct_parameters.rs
@@ -1,7 +1,7 @@
 use validator::Validate;
 
 #[derive(Validate)]
-struct Test<T = ()> {
+struct DefaultParameters<T = ()> {
     a: T,
 }