+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 ## Changelog
 
+## 0.17.1 (unreleased)
+
+- Fix regressions from the derive rewrite
+
+
 ## 0.17.0 (2024/03/04)
 
 - Derive macro has been entirely rewritten
 - Validation is now done through traits that you can implement
-
+- Remove phone validator
+- Remove automatic use of serde rename for fields name (temporary)
 
 ## 0.16.0 (2022/06/27)